Split async
